Fix minus unary operator on unsigned variable

+++ src/cachelot/string_conv.h
@@ -131,7 +131,7 @@ namespace cachelot {
                 UnsignedIntType xvalue;
                 if (negative) {
                     *dest++ = '-';
-                    xvalue = - static_cast<UnsignedIntType>(value);
+                    xvalue = - (long long)static_cast<UnsignedIntType>(value);
                 } else {
                     xvalue = value;
                 }
